// types/CourseRegister.ts
export interface ICandidate {
    code?: string
    name: string
    birthday?: number
    tel: string
    email: string
    registDate?: number
    isDormitory?: boolean
    image?: string
    sex?: number
    identification?: string
    profiles?: []
    statusId?: number
    classId?: number
    majorsId?: number
    submitSuccess?: boolean
    loading?: boolean
}